requirements.txt

paho-mqtt

click